refactor screens into separate files
